// frontend/src/app/models/Owner.ts
export class Owner {
    username: String = "";
    password: String = "";
    firstname: String = "";
    lastname: String = "";
    gender: Number = 0;
    address: String = "";
    phoneNumber: String = "";
    email: String = "";
    picture: String = "";
    creditCardNumber: String = "";
    type: Number = 0;
    activated: boolean = false;
}